backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "channels>=4.3.1",
    "daphne>=4.2.1",
    "dj-database-url>=3.0.1",
    "dj-rest-auth>=7.0.1",
    "django>=5.2.4",
    "django-admin-logs>=1.4.0",
    "django-allauth>=65.11.1",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=23.0.0",
    "paystack>=1.5.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "psycopg2-pool>=1.2",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "whitenoise>=6.9.0",
]
